version=12.2.RC2_dev
# code: enum class for Callbacks 
#   -- NOTE: THIS AFFECTS ALL EMBEDDED APPLICATIONS 
#   -- WITH CALLBACKS (in a good and simple way!)



Former-commit-id: 4a0f7edbbc59aabd6733efc41c4e96551ff6e9f5

+++ Jmol/src/org/openscience/jmol/app/surfacetool/SurfaceTool.java
@@ -363,10 +363,10 @@ private void updateMeshInfo(Shape[] shapes, int kind){
 
   void updateSlices() {
     for (int i = 0; i < surfaceIDs.size(); i++)
-        sliceObject(surfaceIDs.get(i), surfaceKind.get(i));
+        sliceObject(surfaceIDs.get(i), surfaceKind.get(i).intValue());
   }
 
-  void sliceObject(String objectName, Integer kind) {
+  void sliceObject(String objectName, int kind) {
  //   unnecessary already in ID list, which we carefully update
 //    MeshCollection mc = getIsosurfaces(); 
 //    int iMesh = mc.getIndexFromName(objectName);
